// mobile/GreenwichApp/src/resources/types.ts
export type TSignalTypeSuffix = {
  suffix: string;
  signalType: string;
  heading?: string;
};

export type TSignalTypeSuffixHandling = TSignalTypeSuffix & {
  replace?: TSignalTypeSuffix;
};

export type TInputPoint = {
  id: number;
  ioModuleId?: string | number;
  controllerId?: string | number;
  type: string;
  sub_type: string;
  sub_type2?: string;
  name: string;
  mapId: string;
  x: number;
  y: number;
  allowedLocations: string[];
  canonicalName?: string;
  suffix?: string;
  signalType?: string;
  location: string;
};
// ...
export type TRootStackParamList = {
  Home: undefined;
  CCTV: undefined;
  DoorAccess: undefined;
  Electricity: undefined;
  Water: undefined;
  Leakage: undefined;
  AirFlow: undefined;
  WaterTank: undefined;
  Valve: undefined;
  Door: undefined;
  EmergencyAlarm: undefined;
  Lift: undefined;
  Settings: undefined;
};

export type TNavigationProp = NavigationProp<
  TRootStackParamList,
  keyof TRootStackParamList
>;

export type THierarchicalMenu = {
  [type: string]:
    | TInputPoint[]
    | {
        [type: string]: TInputPoint[];
      };
};